mild refactor

+++ src/filters/field.cpp
@@ -441,11 +441,11 @@ void __stdcall SelectRangeEvery::GetAudio(void* buf, __int64 start, __int64 coun
     return;
   }
 
-  int samples_filled = 0;
+  __int64 samples_filled = 0;
   BYTE* samples = (BYTE*)buf;
   const int bps = vi.BytesPerAudioSample();
   int startframe = vi.FramesFromAudioSamples(start);
-  int general_offset = (int)(start - vi.AudioSamplesFromFrames(startframe));  // General compensation for startframe rounding.
+  __int64 general_offset = start - vi.AudioSamplesFromFrames(startframe);  // General compensation for startframe rounding.
 
   while (samples_filled < count) {
     const int iteration = startframe / length;                    // Which iteration is this.
@@ -457,7 +457,7 @@ void __stdcall SelectRangeEvery::GetAudio(void* buf, __int64 start, __int64 coun
     const __int64 start_offset = vi.AudioSamplesFromFrames(iteration * every + iteration_into) + general_offset;
 
     child->GetAudio(&samples[samples_filled*bps], start_offset, getsamples, env);
-    samples_filled += (int)getsamples;
+    samples_filled += getsamples;
     startframe = (iteration+1) * every;
     general_offset = 0; // On the following loops, general offset should be 0, as we are either skipping.
   }
